# v1.5.4 - Introduce a readers system to read logical expressions from text. Also, make logical expression equality more flexible. Makes it easier to add new logical expression objects to hierarchy. - Various minor (but important) optimizations. - Add labeled single sentence data item with lexical entries. - Utility to parse a single sentence with a lexicon. - Support for generalized function composition. - Rewrite how the parser handles unary rules: more correct and better re-use of computations.

+++ mr.lambda.exec.naive/src/edu/uw/cs/lil/tiny/mr/lambda/exec/naive/AbstractEvaluationServices.java
@@ -57,7 +57,7 @@ public abstract class AbstractEvaluationServices<S> implements
 	public Object evaluateConstant(LogicalConstant logicalConstant) {
 		
 		// Try to treat the constant as a number
-		final Integer num = LogicLanguageServices
+		final Long num = LogicLanguageServices
 				.logicalExpressionToInteger(logicalConstant);
 		if (num != null) {
 			return new Double(num);
